Moved the Summary to DetailPath
