+++ frontend/components/lines/LinesDetailHeader/styles.module.css
@@ -34,8 +34,6 @@
 	-webkit-box-orient: vertical;
 }
 
-
-
 /* * */
 /* CONTAINER */
 
@@ -68,61 +66,3 @@
 		margin-top: var(--size-spacing-20);
 	}
 }
-/* CONTAINER */
-
-.containerSummary {
-	position: sticky;
-	top: var(--size-height-header);
-	z-index: 11;
-	display: flex;
-	flex-direction: column;
-	gap: var(--size-spacing-15);
-	align-items: flex-start;
-	justify-content: flex-start;
-	width: 100%;
-	padding: var(--size-spacing-15);
-	background-color: var(--color-system-background-100);
-	border-bottom: 1px solid var(--color-system-border-100);
-	border-top-right-radius: var(--border-radius-lg);
-}
-
-.containerSummary.isSticky {
-	background-color: var(--color-system-background-100);
-	border-radius: unset;
-	box-shadow: 0 0 10px 0 rgb(0 0 0 / 15%);
-}
-
-@media (width < 1000px) {
-	.container {
-		border-radius: unset;
-	}
-}
-
-/* * */
-
-/* LINE SUMMARY WRAPPER */
-
-.linesSummaryWrapper {
-	width: 100%;
-	font-size: 14px;
-	font-weight: var(--font-weight-semibold);
-	color: var(--color-system-text-200);
-}
-
-.linesSummaryWrapper .destinationName,
-.linesSummaryWrapper .dayName {
-	font-size: 15px;
-	font-weight: var(--font-weight-bold);
-	color: var(--color-system-text-100);
-}
-
-.linesSummaryWrapper .changeDay {
-	color: var(--color-system-text-300);
-}
-
-.linesSummaryWrapper .changeDay:hover {
-	color: var(--color-status-info-text);
-	text-decoration: underline;
-}
-
-/* * */
