+++ frontend/components/lines/LinesDetailPath/index.tsx
@@ -8,6 +8,9 @@ import { Surface } from '@/components/layout/Surface';
 import { LinesDetailPathList } from '@/components/lines/LinesDetailPathList';
 import { LinesDetailPathMap } from '@/components/lines/LinesDetailPathMap';
 import { useLinesDetailContext } from '@/contexts/LinesDetail.context';
+import { useOperationalDayContext } from '@/contexts/OperationalDay.context';
+import { useStickyObserver } from '@/hooks/useStickyObserver';
+import { getCssVariableValue } from '@/utils/getCssVariableValue';
 import { useTranslations } from 'next-intl';
 
 import styles from './styles.module.css';
@@ -23,6 +26,9 @@ export function LinesDetailPath() {
 	const t = useTranslations('lines.LinesDetailPath');
 
 	const linesDetailContext = useLinesDetailContext();
+	const operationalDayContext = useOperationalDayContext();
+	const headerHeight = getCssVariableValue('--size-height-header');
+	const { isSticky, ref: stickyElementRef } = useStickyObserver({ top: headerHeight }, [1], { top: -1 });
 
 	//
 	// B. Render components
@@ -37,6 +43,23 @@ export function LinesDetailPath() {
 
 	return (
 		<Surface>
+			<div ref={stickyElementRef} className={`${styles.containerSummary} ${isSticky ? styles.isSticky : ''}`}>
+				{isSticky && (
+					<>
+						<p className={styles.linesSummaryWrapper}>
+							{t.rich('summary', {
+								changeDay: chunks => <a className={styles.changeDay} href="#">{chunks}</a>,
+								day_name: operationalDayContext.data.selected_day_jsdate,
+								dayName: chunks => <span className={styles.dayName}>{chunks}</span>,
+								destination_name: linesDetailContext.data.active_pattern?.headsign,
+								destinationName: chunks => <span className={styles.destinationName}>{chunks}</span>,
+								line_number: linesDetailContext.data.active_pattern?.line_id,
+								lineNumber: chunks => <span className={styles.lineNumber}>{chunks}</span>,
+							})}
+						</p>
+					</>
+				)}
+			</div>
 			<Section>
 				<div className={styles.container}>
 					<LinesDetailPathList />
