+++ frontend/components/lines/LinesDetailPath/styles.module.css
@@ -36,3 +36,61 @@
 		border: none;
 	}
 }
+
+
+/* * */
+/* CONTAINER */
+
+.containerSummary {
+	position: sticky;
+	top: var(--size-height-header);
+	z-index: 11;
+	display: flex;
+	flex-direction: column;
+	gap: var(--size-spacing-15);
+	align-items: flex-start;
+	justify-content: flex-start;
+	width: 100%;
+	background-color: white;
+}
+
+.containerSummary.isSticky {
+	border-bottom: 1px solid var(--color-system-border-100);
+	border-top-right-radius: var(--border-radius-lg);
+	background-color: var(--color-system-background-100);
+	border-radius: unset;
+	box-shadow: 0 0 10px 0 rgb(0 0 0 / 15%);
+	padding: var(--size-spacing-15);
+}
+
+@media (width < 1000px) {
+	.container {
+		border-radius: unset;
+	}
+}
+
+/* * */
+/* LINE SUMMARY WRAPPER */
+
+.linesSummaryWrapper {
+	width: 100%;
+	font-size: 14px;
+	font-weight: var(--font-weight-semibold);
+	color: var(--color-system-text-200);
+}
+
+.linesSummaryWrapper .destinationName, .lineNumber,
+.linesSummaryWrapper .dayName{
+	font-size: 15px;
+	font-weight: var(--font-weight-bold);
+	color: var(--color-system-text-100);
+}
+
+.linesSummaryWrapper .changeDay {
+	color: var(--color-system-text-300);
+}
+
+.linesSummaryWrapper .changeDay:hover {
+	color: var(--color-status-info-text);
+	text-decoration: underline;
+}
